&mut all the clients!
While it's cute that tokio_postgres::Client takes a shared reference
for many methods, it's kind of wrong.  A database connection is
inherently stateful, so trying to share it blindly is not likely
to be right.

Enfore the correct behavior by requiring an exclusive reference
for all our client methods.